update README.md
you may not want to install things by default to /usr/local to
mess around with your system, so the alternative is to configure
it to install to $HOME/.local (using prefix). You should then
make sure "$HOME/.local/bin" is in your path. If not add
  export PATH="$HOME/.local/bin:$PATH"
to your "$HOME/.bashrc" and then open a new terminal.